==> 05_INLA_bym_zscore.Rmd <==
---
title: "05 INLA BYM"
knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'mb_INLA_zscr.html'))})
author: "Simon Brewer"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    theme: flatly
    df_print: paged
    number_sections: true
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Libraries and what not

```{r message=FALSE}
set.seed(42)

library(dplyr)
library(ggpubr)
library(sf)
library(spdep)
library(INLA)
library(INLAutils)
library(tmap)
library(ggregplot)
library(stringr)
```

# Data processing

```{r message=FALSE}
dat_sf <- st_read("./data/glacier_clim.shp")
# dat_sf <- dat_sf[sample(1:nrow(dat_sf), 5000), ]

dat <- dat_sf %>%
  st_drop_geometry() %>% 
  dplyr::select(mb_mwea, area_km2, 
                z_min, z_med, z_max, z_aspct, z_slope, hi,
                east, north, tau,
                t2_w_18, t2_w_b, 
                ppt_a_18, ppt_a_d, ppt_s_18, ppt_s_d) 
# %>%
#   sample_n(5000)
```
### Log-transform

```{r}
dat <- dat %>%
  mutate(area_km2 = log(area_km2),
         z_slope = log(z_slope),
         tau = log(tau),
         ppt_s_18 = log(ppt_s_18)
  )
```

### Scaling

```{r}
dat <- dat %>%
  mutate(area_km2 = scale(area_km2),
         z_med = scale(z_med),
         z_aspct = scale(z_aspct),
         z_slope = scale(z_slope),
         hi = scale(hi),
         tau = scale(tau),
         t2_w_18 = scale(t2_w_18),
         t2_w_b = scale(t2_w_b, center = TRUE),
         ppt_a_18 = scale(ppt_a_18),
         ppt_a_d = scale(ppt_a_d, center = TRUE),
         ppt_s_18 = scale(ppt_s_18),
         ppt_s_d = scale(ppt_s_d, center = TRUE)
  )
```

## INLA spatial graph

```{r}
dat$idarea <- 1:nrow(dat)

# crds <- st_coordinates(dat)
# coo <- cbind(dat$east / 1000,
#              dat$north / 1000)
# 
# nb <- graph2nb(gabrielneigh(coo), sym = TRUE)
# 
# nb2INLA("map.adj", nb)
g <- inla.read.graph("map.adj")
```

# INLA

## Model setup

Priors

```{r}
prior_bym <- list(prec.unstruct=list(prior='pc.prec',param=c(1, 0.025)),
                  prec.spatial=list(prior='pc.prec', param=c(1, 0.025)))

prior_bym2 <- list(
  prec = list(
    prior = "pc.prec",
    param = c(0.01 / 0.31, 0.01)),
  phi = list(
    prior = "pc",
    param = c(0.5, 2 / 3))
)
```

## No interaction

```{r}
f1 <- mb_mwea ~ 
  t2_w_b + t2_w_18 + 
  ppt_a_18 + ppt_a_d + 
  ppt_s_18 + ppt_s_d +
  tau + hi + 
  z_med + z_slope + z_aspct + area_km2 +
  f(idarea, model = "bym2", graph = g, hyper = prior_bym2)

system.time(
  res1 <- inla(f1, data = dat,
               control.predictor = list(compute = TRUE),
               control.compute = list(dic = TRUE, waic = TRUE))
)
```

### Model output

```{r}
summary(res1)
```

Table of coefficients

```{r}
coefs <- round(res1$summary.fixed, 4)
tab_df <- data.frame(variable = row.names(coefs),
                     mean = coefs$mean,
                     CI = paste0("(", coefs$`0.025quant`, ", ",
                                 coefs$`0.975quant`, ")")
                     )
knitr::kable(tab_df)
```

```{r}
Efxplot(res1, Intercept = FALSE) + 
  theme_bw()
```

```{r}
plot_fixed_marginals(res1)
plot_hyper_marginals(res1)
```


### Obs vs pred

```{r}
plot_df <- data.frame(obs = dat$mb_mwea,
                      pred = res1$summary.fitted.values$mean)
ggplot(plot_df, aes(x = obs, y = pred)) +
  geom_point() +
  ggtitle("No interaction model") +
  theme_bw()
```

### Map spatial effect

```{r message = FALSE}
# spatial_re <- res1$summary.random$idarea
# dat_sf$re <- spatial_re$mean[1:nrow(dat_sf)]
# m1 <- tm_shape(dat_sf) +
#   tm_symbols(col = "re", size = 0.25, alpha = 0.75, style = "quantile") 
# print(m1)
```

```{r message = FALSE}
# dat_sf$yhat <- plot_df$pred
# m2 <- tm_shape(dat_sf) +
#   tm_symbols(col = "yhat", size = 0.25, alpha = 0.75, style = "quantile") 
# print(m2)
```

## With interaction

```{r}
f2 <- mb_mwea ~ 
  t2_w_b * t2_w_18 + 
  t2_w_b * ppt_a_18 + ppt_a_d + 
  ppt_s_18 + ppt_s_d +
  hi +
  tau * z_med + z_slope + z_aspct + area_km2 +
  f(idarea, model = "bym2", graph = g, hyper = prior_bym2)

system.time(
  res2 <- inla(f2, data = dat,
               control.predictor = list(compute = TRUE),
               control.compute = list(dic = TRUE, waic = TRUE))
)
```

### Model output

```{r}
summary(res2)
```

```{r}
coefs <- round(res2$summary.fixed, 4)
tab_df <- data.frame(variable = row.names(coefs),
                     mean = coefs$mean,
                     CI = paste0("(", coefs$`0.025quant`, ", ",
                                 coefs$`0.975quant`, ")")
                     )
knitr::kable(tab_df)
```

```{r}
Efxplot(res2, Intercept = FALSE) + 
  theme_bw()
```

```{r}
plot_fixed_marginals(res2)
plot_hyper_marginals(res2)
```


### Obs vs pred

```{r}
plot_df <- data.frame(obs = dat$mb_mwea,
                      pred = res2$summary.fitted.values$mean)
ggplot(plot_df, aes(x = obs, y = pred)) +
  geom_point() +
  ggtitle("Interaction model") +
  theme_bw()
```

## Interaction plots

Extract coefficients

```{r}
coefs <- res2$summary.fixed
```

### tau + z_med

```{r}
b0 <- res2$summary.fixed[rownames(res2$summary.fixed) == "(Intercept)", "mean"]
b1 <- coefs[which(rownames(coefs) == "tau"), "mean"]
x1 <- seq(-3, 3, length = 100) ## Violent crime
b2 <- coefs[which(rownames(coefs) == "z_med"), "mean"]
x2 <- seq(-3, 3) ## PerWhite
b3 <- coefs[which(rownames(coefs) == "tau:z_med"), "mean"]

mydf <- expand.grid(tau = x1, z_med = x2)

mydf$yhat <- b0 + b1 * mydf$tau + b2 * mydf$z_med +
  b3 * (mydf$tau * mydf$z_med)

mydf$z_med <- as.factor(mydf$z_med)

p1 <- ggline(mydf, x = "tau", y = "yhat", 
             col = "z_med", numeric.x.axis = TRUE, 
             size = 1.5, plot_type = 'l',
             xlab = "log(tau)",
             ylab = "mb_mwea", 
             main = "Interaction tau/z_med")
print(p1)
```


### t2_w_b:t2_w_18

```{r}
b0 <- res2$summary.fixed[rownames(res2$summary.fixed) == "(Intercept)", "mean"]
b1 <- coefs[which(rownames(coefs) == "t2_w_b"), "mean"]
x1 <- seq(-3, 3, length = 100) ## Violent crime
b2 <- coefs[which(rownames(coefs) == "t2_w_18"), "mean"]
x2 <- seq(-3, 3) ## PerWhite
b3 <- coefs[which(rownames(coefs) == "t2_w_b:t2_w_18"), "mean"]

mydf <- expand.grid(t2_w_b = x1, t2_w_18 = x2)

mydf$yhat <- b0 + b1 * mydf$t2_w_b + b2 * mydf$t2_w_18 +
  b3 * (mydf$t2_w_b * mydf$t2_w_18)

mydf$t2_w_18 <- as.factor(mydf$t2_w_18)

p2 <- ggline(mydf, x = "t2_w_b", y = "yhat", 
             col = "t2_w_18", numeric.x.axis = TRUE, 
             size = 1.5, plot_type = 'l',
             xlab = "T2 slope",
             ylab = "mb_mwea", 
             main = "Interaction t2_w_b/t2_w_18")
print(p2)
```

```{r}
save(res1, res2, file = "models.RData")
```
