==> 07_gwr.Rmd <==
---
title: "05 GWR"
knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'mb_gwr.html'))})
author: "Simon Brewer"
date: "09/01/2021"
output:
  html_document:
    theme: flatly
    df_print: paged
    number_sections: true
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Libraries and what not

```{r message=FALSE}
set.seed(42)

library(ggplot2)  # for autoplot() generic
library(dplyr)
library(sf)
library(spgwr) ## Maybe change to GWmodel for parallel proc.
library(tmap)
```

# Data processing

```{r}
dat_sf <- st_read("./data/glacier_clim.shp")
dat_sf <- dat_sf[sample(1:nrow(dat_sf), 5000), ]
dat_sp <- as_Spatial(dat_sf)
```

# GWR model

Model formula

```{r}
f1<- mb_mwea ~ t2m_d + tp_d + 
  z_med + z_slope + z_aspct + 
  area_m2 + tau
```

Bandwidth selection

```{r}
bw <- gwr.sel(f1, dat_sp, adapt = TRUE)
print(bw)
```

Run model
```{r}
mb_gwr <- gwr(f1, dat_sp, adapt = bw)
```

```{r}
summary(mb_gwr)
```

# Output

```{r}
tm_shape(mb_gwr$SDF) +
  tm_symbols(col = "localR2", size = 0.25, alpha = 0.75, style = "quantile")
```

```{r}
tm_shape(mb_gwr$SDF) +
  tm_symbols(col = "t2m_d", size = 0.25, alpha = 0.75, style = "quantile")
```

```{r}
tm_shape(mb_gwr$SDF) +
  tm_symbols(col = "tp_d", size = 0.25, alpha = 0.75, style = "quantile")
```

```{r}
tm_shape(mb_gwr$SDF) +
  tm_symbols(col = "z_med", size = 0.25, alpha = 0.75, style = "quantile")
```

```{r}
tm_shape(mb_gwr$SDF) +
  tm_symbols(col = "tau", size = 0.25, alpha = 0.75, style = "quantile")
```
